+++ examples/folders/outputs.tf
@@ -1,12 +1,7 @@
-output "data" {
-  value = [
-    tomap({
-      database = denodo_database_folder.db_folder_ds.database,
-      folder   = denodo_database_folder.db_folder_ds.id
-    }),
-    tomap({
-      database = denodo_database_folder.db_folder_bv.database,
-      folder   = denodo_database_folder.db_folder_bv.id
-    })
-  ]
+output "folder" {
+  value = denodo_database_folder.db_folder.*.id
+}
+
+output "database" {
+  value = denodo_database_folder.db_folder.*.database
 }