+++ examples/jdbc_data_source/variables.tf
@@ -18,51 +18,6 @@ variable "denodo_username" {
   type = string
 }
 
-variable "denodo_database_authentication" {
-  type    = string
-  default = "LOCAL"
-}
-
-variable "denodo_database_char_set" {
-  type    = string
-  default = "DEFAULT"
-}
-
-variable "denodo_database_cost_optimization" {
-  type    = string
-  default = "DEFAULT"
-}
-
-variable "denodo_database_description" {
-  type    = string
-  default = "testing example"
-}
-
-variable "denodo_database_name" {
-  type    = string
-  default = "test_database"
-}
-
-variable "denodo_database_summary_rewrite" {
-  type    = string
-  default = "DEFAULT"
-}
-
-variable "denodo_database_query_simplification" {
-  type    = string
-  default = "DEFAULT"
-}
-
-variable "denodo_dev_user_password" {
-  type      = string
-  sensitive = true
-}
-
-variable "denodo_read_user_password" {
-  type      = string
-  sensitive = true
-}
-
 variable "data_source_catalog_name" {
   type = string
 }