+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [markdownlint](https://dlaa.me/markdownlint/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2021-06-13
+
+### Added to 0.5.0
+
+- Added data source to get metadata from any object in metadata catalog.
+- Added `denodo_dervived_view` resource to create dervived views from files.
+
 ## [0.4.2] - 2021-06-09
 
 ### Added to 0.4.2